Move another test with px to optional
