+++ packages/python/plotly/plotly/tests/test_core/test_figure_messages/test_add_traces.py
@@ -66,36 +66,6 @@ class TestAddTracesMessage(TestCase):
         )
 
 
-class TestAddTracesRowsColsDataTypes(TestCase):
-    def test_add_traces_with_iterable(self):
-        import plotly.express as px
-
-        df = px.data.tips()
-        fig = px.scatter(df, x="total_bill", y="tip", color="day")
-        from plotly.subplots import make_subplots
-
-        fig2 = make_subplots(1, 2)
-        fig2.add_traces(fig.data, rows=[1,] * len(fig.data), cols=[1,] * len(fig.data))
-
-        expected_data_length = 4
-
-        self.assertEqual(expected_data_length, len(fig2.data))
-
-    def test_add_traces_with_integers(self):
-        import plotly.express as px
-
-        df = px.data.tips()
-        fig = px.scatter(df, x="total_bill", y="tip", color="day")
-        from plotly.subplots import make_subplots
-
-        fig2 = make_subplots(1, 2)
-        fig2.add_traces(fig.data, rows=1, cols=2)
-
-        expected_data_length = 4
-
-        self.assertEqual(expected_data_length, len(fig2.data))
-
-
 def test_add_trace_exclude_empty_subplots():
     # Add traces
     fig = make_subplots(2, 2)
